# ICICI_Server/doc/swagger.json
{
    "info": {
        "version": "v0.0.1",
        "title": "ICICI",
        "description": "To perform any operation, you will need to first generate an access token. \n Tokens are generated by invoking the `/auth` API. \n The response is as follows, `Bearer <access-token>`. \n Copy the response and append it to all subsequent APIs where `Authorization` is required."
    },
...
    "paths": {
...
        "/api/product/": {
            "post": {
                "summary": "This endpoint will create product.",
                "description": "A Product is the second part of the Application. It is the catalogue of possessions of a Project. It is the sub-part of Project and can be navigated through Project. A Product like Projects engenders name, version and actions. Name is of the product and its version depicts the times its constructed. The actions include navigating to Flows, navigating to Services , arranges for Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. We can also delete a product through actions. A Project can assimilate only one flow and have multiple multiple Services. \nThis endpoint will create product.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully created.",
                        "schema": {
                            "$ref": "#/responses/createProduct"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "name": "productName",
                        "in": "query",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "description",
                        "in": "query",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "version",
                        "in": "query",
                        "description": "",
                        "required": true
                    },
                    {
                        "name": "file",
                        "in": "formData",
                        "description": "file to upload",
                        "required": true,
                        "type": "file"
                    }
                ],
                "tags": [
                    "Product"
                ]
            },
            "get": {
                "summary": "This endpoint will get product.",
                "description": "A Product is the second part of the Application. It is the catalogue of possessions of a Project. It is the sub-part of Project and can be navigated through Project. A Product like Projects engenders name, version and actions. Name is of the product and its version depicts the times its constructed. The actions include navigating to Flows, navigating to Services , arranges for Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. We can also delete a product through actions. A Project can assimilate only one flow and have multiple multiple Services.\nThis endpoint will get product.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/getProduct"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": false,
                        "description": "",
                        "name": "productId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Product"
                ]
            },
            "put": {
                "summary": "This endpoint will update product.",
                "description": "A Product is the second part of the Application. It is the catalogue of possessions of a Project. It is the sub-part of Project and can be navigated through Project. A Product like Projects engenders name, version and actions. Name is of the product and its version depicts the times its constructed. The actions include navigating to Flows, navigating to Services , arranges for Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. We can also delete a product through actions. A Project can assimilate only one flow and have multiple multiple Services.\nThis endpoint will update product.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "Successfully updated",
                        "schema": {
                            "$ref": "#/responses/updateProduct"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "schema": {
                            "$ref": "#/definitions/updateProduct"
                        },
                        "required": true,
                        "description": "",
                        "name": "body",
                        "in": "body"
                    }
                ],
                "tags": [
                    "Product"
                ]
            },
            "delete": {
                "summary": "This endpoint will delete product.",
                "description": "A Product is the second part of the Application. It is the catalogue of possessions of a Project. It is the sub-part of Project and can be navigated through Project. A Product like Projects engenders name, version and actions. Name is of the product and its version depicts the times its constructed. The actions include navigating to Flows, navigating to Services , arranges for Uploading Data which can be in the form of a WSDL/YAML/JSON file or URL or WSDL/YAML/JSON text. We can also delete a product through actions. A Project can assimilate only one flow and have multiple multiple Services.\nThis endpoint will delete product.",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "schema": {
                            "$ref": "#/responses/deleteProduct"
                        },
                        "headers": {
                            "Access-Control-Allow-Origin": {
                                "type": "string",
                                "description": "URI that may access the resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Validation error"
                    },
                    "500": {
                        "description": "Internal server error."
                    }
                },
                "parameters": [
                    {
                        "required": true,
                        "description": "",
                        "name": "productId",
                        "in": "query"
                    }
                ],
                "tags": [
                    "Product"
                ]
            }
        },
...
    },
...
    "securityDefinitions": {
        "access_token": {
            "description": "Authorization For Transaction APIs",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
    "definitions": {
...
        "updateProduct": {
            "required": [
                "productName",
                "description"
            ],
            "properties": {
                "productId": {
                    "type": "string",
                    "description": "Enter product Id.",
                    "format": "string"
                },
                "productName": {
                    "type": "string",
                    "description": "Enter product name.",
                    "format": "string"
                },
                "description": {
                    "type": "string",
                    "description": "Enter product description.",
                    "format": "string"
                }
            }
        },
...
    },
    "responses": {
...
        "createProduct": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "getProduct": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "updateProduct": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
        "deleteProduct": {
            "required": [],
            "properties": {
                "message": {
                    "type": "string",
                    "description": "You will get message.",
                    "format": "string"
                }
            }
        },
...
    },
...
    "security": [
        {
            "access_token": []
        }
    ],
    "swagger": "2.0"
}